Organize experiment components into separate directory
