+++ lit/src/experiment-components/chat/chat_interface.scss
@@ -1,6 +1,6 @@
-@use "../../../sass/colors";
-@use "../../../sass/common";
-@use "../../../sass/typescale";
+@use "../../sass/colors";
+@use "../../sass/common";
+@use "../../sass/typescale";
 
 :host {
   @include common.flex-column;
