+++ lit/src/experiment-components/chat/chat_interface.ts
@@ -1,7 +1,7 @@
-import "../../../pair-components/button";
-import "../../../pair-components/icon_button";
-import "../../../pair-components/textarea";
-import "../../../pair-components/tooltip";
+import "../../pair-components/button";
+import "../../pair-components/icon_button";
+import "../../pair-components/textarea";
+import "../../pair-components/tooltip";
 import "./chat_message";
 
 import { observable } from "mobx";
@@ -9,10 +9,10 @@ import { MobxLitElement } from "@adobe/lit-mobx";
 import { CSSResultGroup, html } from "lit";
 import { customElement, property } from "lit/decorators.js";
 
-import { core } from "../../../core/core";
-import { ChatService } from "../../../services/chat_service";
+import { core } from "../../core/core";
+import { ChatService } from "../../services/chat_service";
 
-import { ChatMessage } from "../../../shared/types";
+import { ChatMessage } from "../../shared/types";
 
 import { styles } from "./chat_interface.scss";
