+++ lit/src/experiment-components/chat/chat_message.scss
@@ -1,6 +1,6 @@
-@use "../../../sass/colors";
-@use "../../../sass/common";
-@use "../../../sass/typescale";
+@use "../../sass/colors";
+@use "../../sass/common";
+@use "../../sass/typescale";
 
 $avatar-height: 30px;
 $bubble-radius: 20px;
