+++ lit/src/experiment-components/experiment/experiment_config.scss
