+++ lit/src/experiment-components/info/info_config.scss
@@ -1,4 +1,4 @@
-@use "../../../sass/common";
+@use "../../sass/common";
 
 :host {
   @include common.flex-column;
