+++ lit/src/experiment-components/modules/README.md
@@ -1,0 +1,4 @@
+# Experiment Modules
+
+This directory contains experiment modules, where each module is a set
+of associated stages, e.g., for a specific game.
